src/python/pca_processor.py
class PCAProcessor:
    def __init__(self, n_components=0.95):
        self.n_components = n_components
        self.pca = None

    def preprocess_image(self, image):
        # Implement image preprocessing steps such as resizing and normalization
        pass

    def fit(self, dataset):
        # Fit the PCA model to the dataset
        from sklearn.decomposition import PCA
        from sklearn.preprocessing import StandardScaler
        import numpy as np

        # Standardize the dataset
        dataset = np.array(dataset)
        scaler = StandardScaler()
        scaled_data = scaler.fit_transform(dataset)

        # Fit PCA
        self.pca = PCA(n_components=self.n_components)
        self.pca.fit(scaled_data)

    def transform(self, image):
        # Transform the image into PCA space
        processed_image = self.preprocess_image(image)
        return self.pca.transform(processed_image) if self.pca else None

    def inverse_transform(self, pca_data):
        # Inverse transform PCA data back to original space
        return self.pca.inverse_transform(pca_data) if self.pca else None