requirements.txt
Flask
numpy
opencv-python
scikit-learn
Pillow
requests
kagglehub